requirements.txt
streamlit 
opencv-python-headless 
pillow 
numpy 
torch 
torchvision
timm